% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_probs.R
\name{read_probs}
\alias{read_probs}
\title{Read genotype probabilities from database}
\usage{
read_probs(db, chr = NULL, pos = NULL, markers = NULL)
}
\arguments{
\item{db}{Either a filename for a SQLite database, or a connection to such a database.}

\item{chr}{Single chromosome to read.}

\item{pos}{Optional length-2 vector specifying an interval of positions to read.}

\item{markers}{Optional vector of marker names to read. If
\code{pos} is provided, \code{marker} is ignored. The markers
must all be on the same chromosome.}
}
\value{
Genotype probabilities, as an object of class \code{"calc_genoprob"}.
}
\description{
Read genotype probabilities from database
}
