% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_probs.R
\name{write_probs}
\alias{write_probs}
\title{Write genotype probabilities to database}
\usage{
write_probs(db, probs, map = NULL, quiet = FALSE)
}
\arguments{
\item{db}{Either a filename for a SQLite database, or a connection to such a database.}

\item{probs}{Genotype probabilities, as calculated by \code{\link[qtl2geno]{calc_genoprob}}.}

\item{map}{Optional map of marker positions (a list of vectors of positions).}

\item{quiet}{If FALSE, print some information about progress.}
}
\value{
None.
}
\description{
Write genotype probabilities to database
}
